# optimize mkdir calls to avoid base-dir `Mkdir` attempts